+++ src/test/resources/eclim_unit_test_c/src/callhierarchy/main.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include "mod1.h"
+
+int main(void) {
+  printf("Hello World %d", fun1(100));
+  return EXIT_SUCCESS;
+}