+++ src/test/resources/eclim_unit_test_c/src/callhierarchy/mod1.c
@@ -1,0 +1,6 @@
+#include "mod2.h"
+
+int fun1(int n)
+{
+  return fun2(n);
+}