+++ src/test/resources/eclim_unit_test_c/src/callhierarchy/mod1.h
@@ -1,0 +1,6 @@
+#ifndef MOD1_H_
+#define MOD1_H_
+
+int fun1(int n);
+
+#endif /* MOD1_H_ */