+++ src/test/resources/eclim_unit_test_c/src/callhierarchy/mod2.c
@@ -1,0 +1,8 @@
+int fun2(int n) {
+  return n + 1;
+}
+
+int fun3(int n) {
+  fun2(n);
+  return fun1(n) + fun2(n);
+}