+++ src/test/resources/eclim_unit_test_c/src/callhierarchy/mod2.h
@@ -1,0 +1,6 @@
+#ifndef MOD2_H_
+#define MOD2_H_
+
+int fun2(int n);
+
+#endif /* MOD2_H_ */